+++ example/storybook-nativewind/src/components/themed/View/index.tsx
@@ -1,6 +1,0 @@
-import { styled } from '@gluestack-style/react';
-import { View as RNView } from 'react-native';
-
-const StyledRoot = styled(RNView, {}, {});
-
-export const View = StyledRoot;