+++ example/storybook/.storybook/preview.js
@@ -60,7 +60,7 @@ export const parameters = {
           'Feedback',
           ['Alert', 'Progress', 'Spinner', 'Toast'],
           'Data Display',
-          ['Badge', 'Table'],
+          ['Badge', 'Table', 'Card'],
           'Forms',
           [
             'Button',
@@ -123,7 +123,7 @@ export const parameters = {
         'Advanced',
         ['Fonts', 'Animations'],
         'Recipes',
-        ['Card', 'LinearGradient', 'More Recipes'],
+        ['LinearGradient', 'More Recipes'],
         'Resources',
         ['Todo-List', 'Dashboard App', 'Third Party Library Integrations'],
         'Migration',